refactor: keep adapters translation-only, move display decisions to ui
Three presentation decisions lived in adapters:
- xlsx_source rounded Excel's fractional column widths into terminal
  cells, so the domain stored an already-lossy display width. Sheet now
  keeps the width as the file states it (fractional f64); rounding,
  clamping and the NaN guard live next to DEFAULT_CELL_WIDTH in layout.
- terminal_frontend duplicated layout's default-width fallback; it now
  passes the sheet's widths through untouched.
- terminal_frontend composed the editor titles; the dialog now derives
  them in ui from an EditorKind plus the cell address, so wording is a
  ui-only concern.

Snapshots are untouched: the same rounding happens one layer later.

Closes #84

+++ src/domain/sheet.rs
@@ -48,7 +48,7 @@ pub struct Sheet {
     name: String,
     rows: Vec<Vec<CellValue>>,
     /// Author-set column widths from the workbook, index = column.
-    col_widths: Vec<Option<u16>>,
+    col_widths: Vec<Option<f64>>,
     merges: Vec<MergedRange>,
     /// Solid cell backgrounds from the workbook, keyed by (row, col).
     fills: HashMap<(usize, usize), Rgb>,
@@ -97,7 +97,9 @@ impl Sheet {
         self.font_colors.get(&(row, col)).copied()
     }
 
-    pub fn with_col_widths(mut self, widths: Vec<Option<u16>>) -> Self {
+    /// Widths as the file states them: fractional character counts.
+    /// Rounding and clamping into terminal cells is the ui's decision.
+    pub fn with_col_widths(mut self, widths: Vec<Option<f64>>) -> Self {
         self.col_widths = widths;
         self
     }
@@ -182,7 +184,7 @@ impl Sheet {
         self.frozen.1
     }
 
-    pub fn col_width(&self, col: usize) -> Option<u16> {
+    pub fn col_width(&self, col: usize) -> Option<f64> {
         self.col_widths.get(col).copied().flatten()
     }
